requirements.txt
# 核心功能
flask
flask-openapi3
pydantic
werkzeug
gradio
playwright
langchain
langchain-community
langchain-google-genai
pandas
python-dotenv
openai

# AI 模型和代理
smolagents
browser-use
huggingface-hub
google-generativeai
sentence-transformers

# 數據處理
numpy
opencc-python-reimplemented

# 其他工具
requests
beautifulsoup4
pillow
tavily-python
faiss-cpu
google-search-results

# FastAPI
fastapi
uvicorn[standard]